Move tests for Reactions to a separate file
